requirements.txt
PyYAML~=6.0.2
networkx~=3.4.2
pyvis~=0.3.2